Copy autotools files, don't link. Dont't free unallocated mem. Autogen.sh +x
Use --copy arg with libtoolize and automake, so files are copied and not linked.
This way it can be distributed without depending on host files.

Bugfix: program was freeing memory that was not been always allocated with malloc.
Instead, don't free it at all and it will be freed by the OS at program exit.

Make autogen.sh executable