Deleted Quilt's .pc folder
.pc folder is quite equivalent to .git folder for Quilt patches
tool. It keeps the history of applied patches and allows to
revert them, apply again, etc.

Changes are being registered now with Git and not with Quilt,
so this .pc folder is kind "out of sync". Changes commited to
Git will eventually make not possible to revert the patches
due to the changes in the source, so .pc folder will be useless.